monitor/requirements.txt
psutil==5.9.8
requests==2.32.3
APScheduler==3.10.4
python-dotenv==1.0.1
